+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 * fix syntax-rules macro that manipulate code (see tests/syntax.scm and undswap macro)
 * fix `(read)` in the Web REPL
 * fix Node REPL with `npm install` [#305](https://github.com/jcubic/lips/issues/305)
+* fix formatting `syntax-rules`
 
 ## 1.0.0-beta.18
 ### Breaking